Change C.ulong type to C.size_t

+++ radamsa.go
@@ -80,10 +80,10 @@ func (r *Radamsa) Fuzz(input []byte, length int, output []byte, capacity int) (u
 
 func fuzz(input []byte, length int, output []byte, capacity int, seed int64) uint {
 	n := C.radamsa(
-		(*C.uchar)(unsafe.Pointer(&input[0])),  // ptr
-		C.ulong(uint(length)),                  // len
+		(*C.uchar)(unsafe.Pointer(&input[0])),	// ptr
+		C.size_t(uint(length)),									// len
 		(*C.uchar)(unsafe.Pointer(&output[0])), // target
-		C.ulong(uint(capacity)),                // max
+		C.size_t(uint(capacity)),								// max
 		C.uint(uint(seed)))                     // seed
 	return uint(n)
 }
